Created common folder and added common functions from faqs
