+++ resources/faqs.py
@@ -1,27 +1,9 @@
 from flask_restful import Resource
 from flask import request,jsonify,Response
 from sqlalchemy.orm.exc import NoResultFound
-from app import error
+from common.utils import error,headers,is_logged_in,has_admin_privileges
 from models.faqs import Faqs
-from models.users import Users
 
-def is_logged_in():
-    user_token=request.headers.get("X-WWW-USER-TOKEN")
-    try:
-        user=Users.query.filter(Users.__table__.c.auth_token==user_token).one()
-        return user
-    except:
-        return False
-
-def has_admin_privileges():
-    user=is_logged_in()
-    if user:
-        if user.role<9:
-            return True
-        else:
-            return False
-    else:
-        return "Not logged in"
 class Faqs_RUD(Resource):
     """
     For GET PUT and DELETE for specific faq
@@ -29,7 +11,7 @@ class Faqs_RUD(Resource):
     """
     def get(self,faq_id):
         try:
-            faq_object=Faqs.query.filter(Faqs.__table__.c.id == faq_id).one()
+            faq_object = Faqs.query.filter(Faqs.__table__.c.id == faq_id).one()
             ret={
                     "id":faq_object.id,
                     "priority":faq_object.priority,
@@ -39,18 +21,18 @@ class Faqs_RUD(Resource):
                     "placement":faq_object.placement,
                     "user":faq_object.user_id
                 }
-            return (ret,200,{"X-XSS-Protection":"1"})
+            return (ret,200,headers)
 
         except NoResultFound:
-            return (error,404,{"X-XSS-Protection":"1"})
+            return (error,404,headers)
 
 
     def put(self,faq_id):
-        user_status=has_admin_privileges()
-        if user_status=="Not logged in":
-            if user_status==True:
+        user_statu = has_admin_privileges()
+        if user_status == "Not logged in":
+            if user_status == True:
                 try:
-                    faq_object=Faqs.query.get(faq_id)
+                    faq_object = Faqs.query.get(faq_id)
                     return faq_object.id
                 except:
                     return "Something failed"
